refactor: remove JavaScriptRunner dependency from ActionHandler and SupervisorBuilder
